pyproject.toml
[project]
name = "linktunnel"
version = "0.8.4"
description = "Serial bridge, TCP relay, Mihomo local proxy, Grbl helpers, Clash API, discovery"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyserial>=3.5",
  "click>=8.1",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
ble = ["bleak>=0.21"]
i2c = ["smbus2>=0.4"]
dev = ["pytest>=7.4", "pytest-asyncio>=0.23", "pytest-qt>=4.2", "ruff>=0.4"]
desktop = ["pywebview>=5.0"]
gui = ["PyQt6>=6.6"]
gui-full = ["linktunnel[gui,ble,i2c,desktop]"]

[project.scripts]
linktunnel = "linktunnel.cli:main"
linktunnel-gui = "linktunnel.desktop_gui:gui_main"
linktunnel-client = "linktunnel.client_app:client_main"
linktunnel-unified = "linktunnel.unified_gui.__main__:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = [
  "E501",  # line length (Click help strings)
]
